Fixed refactoring
